Refactor the App implementation block by separating concerns
