+++ src/cli/cmd/tui/app.rs
@@ -1,12 +1,5 @@
 // Library
-use crossterm::event::{self, Event, KeyCode, KeyEvent, KeyEventKind};
-use ratatui::layout::{Alignment, Constraint, Direction, Layout};
-use ratatui::style::{Color, Modifier, Style};
-use ratatui::text::{Line, Span};
-use ratatui::widgets::{Block, Borders, Paragraph};
-use ratatui::{DefaultTerminal, Frame};
-
-use crate::utils::helpers;
+use ratatui::DefaultTerminal;
 
 use super::row::Row;
 use super::View;
@@ -14,11 +7,11 @@ use super::View;
 /// The main application state
 #[derive(Debug, Default)]
 pub struct App {
-    cfg: View,          // Configuration parameters
-    data: Vec<Row>,     // The 2-D vector of data
-    total_bytes: usize, // The total count of bytes
-    selected: usize,    // The index of the selected byte
-    exit: bool,         // Should exit the application
+    pub cfg: View,          // Configuration parameters
+    pub data: Vec<Row>,     // The 2-D vector of data
+    pub total_bytes: usize, // The total count of bytes
+    pub selected: usize,    // The index of the selected byte
+    pub exit: bool,         // Should exit the application
 }
 
 impl App {
@@ -84,166 +77,5 @@ impl App {
             self.handle_events()?;
         }
         Ok(())
-    }
-
-    // --
-    // UI
-    // --
-
-    fn draw(&self, f: &mut Frame) {
-        // Create the base layout
-        let base_layout = Layout::default()
-            .direction(Direction::Vertical)
-            .constraints([Constraint::Length(self.data.len() as u16 + 2)].as_ref())
-            .split(f.area());
-
-        let offset_len = 8 + 4;
-        let hex_len = ((self.cfg.format.size() + 1) * self.cfg.size)
-            + (self.cfg.size / self.cfg.group_size)
-            + 2;
-        let ascii_len = (self.cfg.size + 1) + (self.cfg.size / self.cfg.group_size) + 2;
-
-        // Create a layout with three vertical sections
-        let columns = Layout::default()
-            .direction(Direction::Horizontal)
-            .constraints(
-                [
-                    Constraint::Length(offset_len),       // Offset
-                    Constraint::Length(hex_len as u16),   // Hex Values
-                    Constraint::Length(ascii_len as u16), // ASCII Values
-                ]
-                .as_ref(),
-            )
-            .split(base_layout[0]);
-
-        // Create a block with borders and title for each column
-        let offset_block = Block::default().borders(Borders::ALL);
-        let hex_block = Block::default().borders(Borders::ALL);
-        let ascii_block = Block::default().borders(Borders::ALL);
-
-        let mut offset_data = Vec::new();
-        let mut hex_data = Vec::new();
-        let mut ascii_data = Vec::new();
-
-        let selected_styles = Style::default().bg(Color::Yellow);
-        let regular_styles = Style::default().fg(Color::White);
-
-        for (i, row) in self.data.iter().enumerate() {
-            // Offset column
-            offset_data.push(row.format_offset());
-
-            // Hex Values column
-            let mut hex_spans = Vec::new();
-            let mut ascii_spans = Vec::new();
-            for (j, byte) in row.data.iter().enumerate() {
-                // Group values by applying spacing
-                if j > 0 && j % self.cfg.group_size == 0 {
-                    hex_spans.push(Span::from(" "));
-                    ascii_spans.push(Span::from(" "));
-                }
-
-                let byte_str = self.cfg.format.format(*byte);
-                let ascii_str = if helpers::is_printable_ascii_character(byte) {
-                    (*byte as char).to_string()
-                } else {
-                    String::from("·")
-                };
-
-                let style = if i * self.cfg.size + j == self.selected {
-                    selected_styles
-                } else {
-                    regular_styles
-                };
-                hex_spans.push(Span::styled(byte_str, style));
-                hex_spans.push(Span::from(" "));
-                ascii_spans.push(Span::styled(ascii_str, style));
-            }
-
-            hex_data.push(Line::from(hex_spans));
-            ascii_data.push(Line::from(ascii_spans));
-        }
-
-        let offset_paragraph = Paragraph::new(offset_data)
-            .block(offset_block)
-            .alignment(Alignment::Center)
-            .style(Style::default().fg(Color::White));
-        let hex_paragraph = Paragraph::new(hex_data)
-            .block(hex_block)
-            .alignment(Alignment::Center)
-            .style(Style::default().fg(Color::White));
-        let ascii_paragraph = Paragraph::new(ascii_data)
-            .block(ascii_block)
-            .alignment(Alignment::Center)
-            .style(Style::default().fg(Color::White));
-
-        f.render_widget(offset_paragraph, columns[0]);
-        f.render_widget(hex_paragraph, columns[1]);
-        f.render_widget(ascii_paragraph, columns[2]);
-    }
-
-    // --------------
-    // EVENT HANDLERS
-    // --------------
-
-    /// updates the application's state based on user input
-    fn handle_events(&mut self) -> std::io::Result<()> {
-        match event::read()? {
-            // it's important to check that the event is a key press event as
-            // crossterm also emits key release and repeat events on Windows.
-            Event::Key(key_event) if key_event.kind == KeyEventKind::Press => {
-                self.handle_key_event(key_event)
-            }
-            _ => {}
-        }
-        Ok(())
-    }
-
-    /// matches the given key-event and calls the corresponding handler
-    fn handle_key_event(&mut self, key_event: KeyEvent) {
-        match key_event.code {
-            KeyCode::Char('q') => self.exit(),
-            KeyCode::Up => self.move_selection_up(),
-            KeyCode::Right => self.move_selection_left(),
-            KeyCode::Down => self.move_selection_down(),
-            KeyCode::Left => self.move_selection_right(),
-            _ => {}
-        }
-    }
-
-    // ----------------
-    // COMMAND HANDLERS
-    // ----------------
-
-    // Select the element in the row above
-    fn move_selection_up(&mut self) {
-        if self.selected > self.cfg.size {
-            self.selected -= self.cfg.size;
-        }
-    }
-
-    /// Select the previous element
-    fn move_selection_right(&mut self) {
-        if self.selected > 0 {
-            self.selected -= 1;
-        }
-    }
-
-    // Select the element in the row above
-    fn move_selection_down(&mut self) {
-        if self.selected + self.cfg.size < self.total_bytes {
-            self.selected += self.cfg.size;
-        }
-    }
-
-    /// Select the next element
-    fn move_selection_left(&mut self) {
-        if self.selected < self.total_bytes - 1 {
-            self.selected += 1;
-        }
-    }
-
-    /// Exits the application
-    fn exit(&mut self) {
-        self.exit = true;
     }
 }
